requirements.txt
absolufy-imports==0.3.1
accelerate==1.10.1
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
authlib==1.6.4
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
cryptography==46.0.1
dill==0.4.0
docling==2.54.0
docling-core==2.48.2
docling-ibm-models==3.9.1
docling-parse==4.5.0
docstring-parser==0.17.0
dotenv==0.9.9
easyocr==1.7.2
et-xmlfile==2.0.0
faker==37.8.0
fastapi==0.117.1
filelock==3.19.1
filetype==1.2.0
fsspec==2025.9.0
google-adk==1.14.1
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.115.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.58.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.38.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
idna==3.10
imageio==2.37.0
importlib-metadata==8.7.0
jinja2==3.1.6
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
lazy-loader==0.4
lxml==5.4.0
mako==1.3.10
markdown-it-py==4.0.0
marko==2.2.0
markupsafe==3.0.2
mcp==1.14.1
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.5
ninja==1.13.0
numpy==2.3.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pluggy==1.6.0
polyfactory==2.22.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pylatexenc==2.10
pyparsing==3.2.5
pypdf2==3.0.1
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tifffile==2025.9.20
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.2
typer==0.16.1
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
websockets==15.0.1
xlsxwriter==3.2.9
zipp==3.23.0
